# docs/swagger.yaml
basePath: /service/api
definitions:
...
  dto.BaseResponse:
    properties:
      data: {}
      message_en:
        example: Success
        type: string
      message_th:
        example: สำเร็จ
        type: string
      status:
        example: success
        type: string
      status_code:
        example: 200
        type: integer
    type: object
...
  dto.CreateExpenseDTO:
    properties:
      amount:
        type: number
      bank_id:
        type: string
      currency:
        type: string
      description:
        type: string
      note:
        type: string
      payment_method:
        type: string
      reference_no:
        type: string
      transaction_category_id:
        type: string
      txn_date:
        type: string
    required:
    - amount
    - bank_id
    - currency
    - description
    - transaction_category_id
    - txn_date
    type: object
...
  dto.ExpenseDTO:
    properties:
      amount:
        type: number
      bank_id:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      currency:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      expense_id:
        type: string
      note:
        type: string
      payment_method:
        type: string
      reference_no:
        type: string
      transaction_category_id:
        type: string
      txn_date:
        type: string
      updated_at:
        type: string
    type: object
  dto.ExpenseSummaryDTO:
    properties:
      total_all:
        type: number
      total_this_month:
        type: number
      total_today:
        type: number
    type: object
...
  dto.Pagination:
    properties:
      list:
        items: {}
        type: array
      page:
        example: 1
        type: integer
      size:
        example: 10
        type: integer
      total_count:
        example: 100
        type: integer
      total_pages:
        example: 10
        type: integer
    type: object
...
  dto.UpdateExpenseDTO:
    properties:
      amount:
        type: number
      bank_id:
        type: string
      currency:
        type: string
      description:
        type: string
      note:
        type: string
      payment_method:
        type: string
      reference_no:
        type: string
      transaction_category_id:
        type: string
      txn_date:
        type: string
    required:
    - amount
    - bank_id
    - currency
    - description
    - transaction_category_id
    - txn_date
    type: object
...
host: api.dev.rkp-media.com
info:
  contact: {}
  description: This is an ERP API demo.
  title: ERP Demo API
  version: "1.0"
paths:
...
  /v1/expense/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete an expense by its ID
      parameters:
      - description: Expense ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      summary: Delete Expense by ID
      tags:
      - Expense
    get:
      consumes:
      - application/json
      description: Get a single expense by its ID
      parameters:
      - description: Expense ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.ExpenseDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      summary: Get Expense by ID
      tags:
      - Expense
    put:
      consumes:
      - application/json
      description: Update an existing expense by its ID
      parameters:
      - description: Expense ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Expense
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateExpenseDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      summary: Update Expense by ID
      tags:
      - Expense
  /v1/expense/create:
    post:
      consumes:
      - application/json
      description: Create a new expense
      parameters:
      - description: Create Expense
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateExpenseDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.ExpenseDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      summary: Create Expense
      tags:
      - Expense
  /v1/expense/list:
    get:
      consumes:
      - application/json
      description: List all expenses with pagination and optional search
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        maximum: 100
        name: limit
        type: integer
      - description: Search term
        in: query
        name: search
        type: string
      - default: '"created_at"'
        description: Field to sort by
        in: query
        name: sortBy
        type: string
      - default: '"desc"'
        description: Sort order
        enum:
        - asc
        - desc
        in: query
        name: sortOrder
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.Pagination'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      summary: List Expenses
      tags:
      - Expense
  /v1/expense/summary:
    get:
      consumes:
      - application/json
      description: Get summary of expense for today, this month, and all time
      parameters:
      - description: Bank ID to filter expenses
        in: query
        name: bank_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.ExpenseSummaryDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      summary: Expense Summary by Filter
      tags:
      - Expense
...
swagger: "2.0"